+++ Puzzle/graphicsclass.h
@@ -41,7 +41,7 @@ public:
 
 	bool Initialize(int, int, HWND);
 	void Shutdown();
-	bool Frame(int, int, float);
+	bool Frame(float, int, float);
 
 private:
 	bool Render();
